+++ modules/nodes.py
@@ -2,7 +2,7 @@ import os
 
 import torch
 import numpy as np
-from .utils import *
+from .video_utils import *
 
 import folder_paths
